Payments/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.0
fastapi-limiter==0.1.6
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
limits==3.13.0
Mako==1.3.5
MarkupSafe==2.1.5
mercadopago==2.2.3
packaging==24.1
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.3
pytest-mock==3.14.0
python-dotenv==1.0.1
python-multipart==0.0.12
redis==5.2.0
requests==2.32.3
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
wrapt==1.16.0
